# apps/backend/tests/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "Body_upload_resume_api_v1_resume_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_resume_upload_post",
        "type": "object"
      },
      "Body_upload_resume_api_v1_resumes_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_api_v1_resumes_upload_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeImprovementRequest": {
        "properties": {
          "job_id": {
            "description": "DB UUID reference to the job",
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "resume_id": {
            "description": "DB UUID reference to the resume",
            "format": "uuid",
            "title": "Resume Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "resume_id"
        ],
        "title": "ResumeImprovementRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Resume Matcher",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/resume": {
      "get": {
        "description": "Retrieves resume data from both resume_model and processed_resume model by resume_id.\n\nArgs:\n    resume_id: The ID of the resume to retrieve\n\nReturns:\n    Combined data from both resume and processed_resume models\n\nRaises:\n    HTTPException: If the resume is not found or if there's an error fetching data.",
        "operationId": "get_resume_api_v1_resume_get",
        "parameters": [
          {
            "description": "Resume ID to fetch data for",
            "in": "query",
            "name": "resume_id",
            "required": true,
            "schema": {
              "description": "Resume ID to fetch data for",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get resume data from both resume and processed_resume models",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/resume/improve": {
      "post": {
        "description": "Scores and improves a resume against a job description.\n\nRaises:\n    HTTPException: If the resume or job is not found.",
        "operationId": "score_and_improve_api_v1_resume_improve_post",
        "parameters": [
          {
            "description": "Enable streaming response using Server-Sent Events",
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": false,
              "description": "Enable streaming response using Server-Sent Events",
              "title": "Stream",
              "type": "boolean"
            }
          },
          {
            "description": "If false, only deterministic baseline improvement is applied (no LLM call)",
            "in": "query",
            "name": "use_llm",
            "required": false,
            "schema": {
              "default": true,
              "description": "If false, only deterministic baseline improvement is applied (no LLM call)",
              "title": "Use Llm",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeImprovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score and improve a resume against a job description",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/resume/upload": {
      "post": {
        "description": "Accepts a PDF or DOCX file, converts it to HTML/Markdown, and stores it in the database.\n\nRaises:\n    HTTPException: If the file type is not supported or if the file is empty.",
        "operationId": "upload_resume_api_v1_resume_upload_post",
        "parameters": [
          {
            "description": "Defer structured extraction (faster response, background processing)",
            "in": "query",
            "name": "defer",
            "required": false,
            "schema": {
              "default": false,
              "description": "Defer structured extraction (faster response, background processing)",
              "title": "Defer",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_resume_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a resume in PDF or DOCX format and store it into DB in HTML/Markdown format",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/resumes": {
      "get": {
        "description": "Retrieves resume data from both resume_model and processed_resume model by resume_id.\n\nArgs:\n    resume_id: The ID of the resume to retrieve\n\nReturns:\n    Combined data from both resume and processed_resume models\n\nRaises:\n    HTTPException: If the resume is not found or if there's an error fetching data.",
        "operationId": "get_resume_api_v1_resumes_get",
        "parameters": [
          {
            "description": "Resume ID to fetch data for",
            "in": "query",
            "name": "resume_id",
            "required": true,
            "schema": {
              "description": "Resume ID to fetch data for",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get resume data from both resume and processed_resume models",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/resumes/improve": {
      "post": {
        "description": "Scores and improves a resume against a job description.\n\nRaises:\n    HTTPException: If the resume or job is not found.",
        "operationId": "score_and_improve_api_v1_resumes_improve_post",
        "parameters": [
          {
            "description": "Enable streaming response using Server-Sent Events",
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": false,
              "description": "Enable streaming response using Server-Sent Events",
              "title": "Stream",
              "type": "boolean"
            }
          },
          {
            "description": "If false, only deterministic baseline improvement is applied (no LLM call)",
            "in": "query",
            "name": "use_llm",
            "required": false,
            "schema": {
              "default": true,
              "description": "If false, only deterministic baseline improvement is applied (no LLM call)",
              "title": "Use Llm",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeImprovementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score and improve a resume against a job description",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/resumes/upload": {
      "post": {
        "description": "Accepts a PDF or DOCX file, converts it to HTML/Markdown, and stores it in the database.\n\nRaises:\n    HTTPException: If the file type is not supported or if the file is empty.",
        "operationId": "upload_resume_api_v1_resumes_upload_post",
        "parameters": [
          {
            "description": "Defer structured extraction (faster response, background processing)",
            "in": "query",
            "name": "defer",
            "required": false,
            "schema": {
              "default": false,
              "description": "Defer structured extraction (faster response, background processing)",
              "title": "Defer",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_resumes_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a resume in PDF or DOCX format and store it into DB in HTML/Markdown format",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}